Moved files to work with automated tests
